split bbc into its own library to allow us in g4hough without circular dependency
