+++ simulation/g4simulation/g4bbc/BbcVertexFastSimReco.C
