+++ simulation/g4simulation/g4bbc/Makefile.am
@@ -6,8 +6,8 @@ INCLUDES = \
   -I`root-config --incdir`
 
 lib_LTLIBRARIES = \
-   libg4vertex_io.la \
-   libg4vertex.la
+   libg4bbc_io.la \
+   libg4bbc.la
 
 AM_CXXFLAGS = \
  -Wall -Werror -msse2
@@ -16,41 +16,37 @@ AM_LDFLAGS = \
   -L$(libdir) \
   -L$(OFFLINE_MAIN)/lib
 
-libg4vertex_io_la_LDFLAGS = \
+libg4bbc_io_la_LDFLAGS = \
   -L$(OFFLINE_MAIN)/lib
 
-libg4vertex_io_la_LIBADD = \
+libg4bbc_io_la_LIBADD = \
   -lphool
 
-libg4vertex_la_LIBADD = \
+libg4bbc_la_LIBADD = \
   -lg4detectors \
   -lfun4all \
-  -lg4bbc_io \
-  -lg4hough_io \
-  libg4vertex_io.la
+  libg4bbc_io.la
 
 pkginclude_HEADERS = \
-  GlobalVertex.h \
-  GlobalVertex_v1.h \
-  GlobalVertexMap.h \
-  GlobalVertexMap_v1.h \
-  GlobalVertexReco.h
+  BbcVertex.h \
+  BbcVertex_v1.h \
+  BbcVertexMap.h \
+  BbcVertexMap_v1.h \
+  BbcVertexFastSimReco.h
 
-libg4vertex_io_la_SOURCES = \
-  GlobalVertex.C \
-  GlobalVertex_Dict.C \
-  GlobalVertex_v1.C \
-  GlobalVertex_v1_Dict.C \
-  GlobalVertexMap.C \
-  GlobalVertexMap_Dict.C \
-  GlobalVertexMap_v1.C \
-  GlobalVertexMap_v1_Dict.C
+libg4bbc_io_la_SOURCES = \
+  BbcVertex.C \
+  BbcVertex_Dict.C \
+  BbcVertex_v1.C \
+  BbcVertex_v1_Dict.C \
+  BbcVertexMap.C \
+  BbcVertexMap_Dict.C \
+  BbcVertexMap_v1.C \
+  BbcVertexMap_v1_Dict.C
 
-libg4vertex_la_SOURCES = \
-  GlobalVertexFastSimReco.C \
-  GlobalVertexFastSimReco_Dict.C \
-  GlobalVertexReco.C \
-  GlobalVertexReco_Dict.C
+libg4bbc_la_SOURCES = \
+  BbcVertexFastSimReco.C \
+  BbcVertexFastSimReco_Dict.C 
 
 # Rule for generating table CINT dictionaries.
 %_Dict.C: %.h %LinkDef.h
@@ -68,8 +64,8 @@ BUILT_SOURCES = \
   testexternals.C
 
 testexternals_LDADD = \
-  libg4vertex_io.la \
-  libg4vertex.la
+  libg4bbc_io.la \
+  libg4bbc.la
 
 testexternals.C:
 	echo "//*** this is a generated file. Do not commit, do not edit" > $@
