+++ simulation/g4simulation/g4bbc/configure.in
@@ -1,0 +1,16 @@
+AC_INIT(configure.in)
+
+AM_INIT_AUTOMAKE(g4bbc, 1.00)
+
+AC_PROG_CXX(g++)
+AC_ENABLE_STATIC(no)
+LT_INIT
+
+dnl   no point in suppressing warnings people should 
+dnl   at least see them, so here we go for g++: -Wall
+if test $ac_cv_prog_gxx = yes; then
+   CXXFLAGS="$CXXFLAGS -Wall -Werror"
+fi
+
+AC_OUTPUT(Makefile)
+
