+++ simulation/g4simulation/g4hough/Makefile.am
@@ -31,7 +31,8 @@ libg4hough_la_LIBADD = \
   -lHelixHough \
   -lFitNewton \
   -lSeamstress \
-  -lcemc \
+  -lg4bbc_io \
+  -lcemc_io \
   libg4hough_io.la
 
 pkginclude_HEADERS = \
